Update check vignettes
