--- R_ignore/inference_checks/source/doublewell_loglike_logprior.cpp
@@ -1,31 +1,0 @@
-#include <Rcpp.h>
-using namespace Rcpp;
-
-// [[Rcpp::export]]
-SEXP loglike(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc) {
-  double mu = params["mu"];
-  double gamma = params["gamma"];
-  double ret = -gamma*(mu*mu - 1.0)*(mu*mu - 1.0);
-  
-  return Rcpp::wrap(ret);
-}
-
-// [[Rcpp::export]]
-SEXP logprior(Rcpp::NumericVector params, Rcpp::List misc) {
-  return Rcpp::wrap(0.0);
-}
-
-// [[Rcpp::export]]  
-SEXP create_xptr(std::string function_name) {  
-  typedef SEXP (*funcPtr_likelihood)(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc);  
-  typedef SEXP (*funcPtr_prior)(Rcpp::NumericVector params, Rcpp::List misc);  
-  
-  if (function_name == "loglike"){
-    return(Rcpp::XPtr<funcPtr_likelihood>(new funcPtr_likelihood(&loglike)));
-  } 
-  if (function_name == "logprior"){
-    return(Rcpp::XPtr<funcPtr_prior>(new funcPtr_prior(&logprior)));
-  } 
-  
-  stop("cpp function %i not found", function_name);
-}
