--- R_ignore/inference_checks/source/returnprior_loglike_logprior.cpp
@@ -1,33 +1,0 @@
-#include <Rcpp.h>
-using namespace Rcpp;
-
-// [[Rcpp::export]]
-SEXP loglike(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc) {
-  return Rcpp::wrap(0.0);
-}
-
-// [[Rcpp::export]]
-SEXP logprior(Rcpp::NumericVector params, Rcpp::List misc) {
-  
-  double ret = R::dnorm(params["real_line"], 0.0, 1.0, true) +
-    R::dgamma(-params["neg_line"], 5.0, 5.0, true) +
-    R::dgamma(params["pos_line"], 5.0, 5.0, true) +
-    R::dbeta(params["unit_interval"], 3.0, 3.0, true);
-  
-  return Rcpp::wrap(ret);
-}
-
-// [[Rcpp::export]]  
-SEXP create_xptr(std::string function_name) {  
-  typedef SEXP (*funcPtr_likelihood)(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc);  
-  typedef SEXP (*funcPtr_prior)(Rcpp::NumericVector params, Rcpp::List misc);  
-  
-  if (function_name == "loglike"){
-    return(Rcpp::XPtr<funcPtr_likelihood>(new funcPtr_likelihood(&loglike)));
-  } 
-  if (function_name == "logprior"){
-    return(Rcpp::XPtr<funcPtr_prior>(new funcPtr_prior(&logprior)));
-  } 
-  
-  stop("cpp function %i not found", function_name);
-}
