+++ inst/extdata/checks/normal_loglike_logprior.cpp
@@ -20,21 +20,19 @@ SEXP logprior(Rcpp::NumericVector params, Rcpp::List misc) {
 }
 
 // [[Rcpp::export]]  
-SEXP create_xptr_loglike(std::string fstr) {  
-  typedef SEXP (*funcPtr)(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc) ;  
-  if (fstr == "loglike"){
-    return(Rcpp::XPtr<funcPtr>(new funcPtr(&loglike))) ;
-  } else {
-    return(Rcpp::XPtr<funcPtr>(R_NilValue)) ; 
-  }
+SEXP create_xptr(std::string function_name) {  
+  typedef SEXP (*funcPtr_likelihood)(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc);  
+  typedef SEXP (*funcPtr_prior)(Rcpp::NumericVector params, Rcpp::List misc);  
+  
+  if (function_name == "loglike"){
+    return(Rcpp::XPtr<funcPtr_likelihood>(new funcPtr_likelihood(&loglike)));
+  } 
+  
+  if (function_name == "logprior"){
+    return(Rcpp::XPtr<funcPtr_prior>(new funcPtr_prior(&logprior)));
+  } 
+
+  stop("cpp function %i not found", function_name);
 }
 
-// [[Rcpp::export]]  
-SEXP create_xptr_logprior(std::string fstr) {  
-  typedef SEXP (*funcPtr)(Rcpp::NumericVector params, Rcpp::List misc) ;  
-  if (fstr == "logprior"){
-    return(Rcpp::XPtr<funcPtr>(new funcPtr(&logprior))) ;
-  } else {
-    return(Rcpp::XPtr<funcPtr>(R_NilValue)) ; 
-  }
-}  
+
