+++ vignettes/checks_normal_model.Rmd
@@ -21,7 +21,7 @@ set.seed(1)
 library(drjacoby)
 ```
 
-**Purpose:** to compare *drjacoby* infered posterior distribution against the exact analytical solution for a very simple model of Normally distributed data.
+**Purpose:** to compare *drjacoby* infered posterior distribution against the exact analytical solution for a very simple model.
 
 ## Model
 
@@ -34,9 +34,8 @@ data_list <- list(x = rnorm(n))
 
 Likelihood and prior:
 
-```{r}
-cpp_funcs <- system.file("extdata/checks/", "normal_loglike_logprior.cpp", package = 'drjacoby', mustWork = TRUE)
-Rcpp::sourceCpp(cpp_funcs)
+```{r, echo = FALSE, comment = ''}
+Rcpp::sourceCpp(system.file("extdata/checks/", "normal_loglike_logprior.cpp", package = 'drjacoby', mustWork = TRUE))
 ```
 
 Parameters dataframe:
@@ -48,7 +47,6 @@ df_params <- define_params(name = "mu", min = -1, max = 1)
 ## MCMC
 
 ```{r cars}
-t0 <- Sys.time()
 mcmc <- run_mcmc(data = data_list,
                  df_params = df_params,
                  loglike = "loglike",
@@ -56,8 +54,7 @@ mcmc <- run_mcmc(data = data_list,
                  burnin = 1e3,
                  samples = 1e5,
                  chains = 10,
-                 pb_markdown = TRUE)
-Sys.time() - t0
+                 silent = TRUE)
 ```
 
 ## Posterior plots
@@ -74,4 +71,3 @@ fx <- dnorm(x, mean = mean(data_list$x), sd = sqrt(1/n))
 hist(mu_draws, breaks = seq(-1, 1, 0.01), probability = TRUE, col = "black")
 lines(x, fx, col = 2, lwd = 4)
 ```
-
